Added post method to ApiDataset
This post method uses the GetId function to (hopefully) add the dataset
to the database. Known issue of retrieving the dataset. I don't know if
it is caused by this function or the GetId of DbComment.
Also this method does not interact with the analysis layer yet. That
functionality needs to be added.