Add code to fix DatasetAnalysis and ApiDataset
Fixed ApiDataset error where recordValidate was not defined. Fixed by defining
it.
Added a single if statement to DatasetAnalysis. If statement check whether a
topic model id for a comment is not -1. A topic model id can be -1 when the
topic model analysis does not assign a topic to a comment. We are leaving these
comments out of the results.